+++ apps/portal/src/components/code/code.client.tsx
@@ -1,6 +1,3 @@
 "use client";
 
-export {
-  CodeClient,
-  type CodeProps,
-} from "@workspace/ui/components/code/code.client";
+export { CodeClient } from "@workspace/ui/components/code/code.client";
