--- apps/portal/src/components/code/plaintext-code.tsx
@@ -1,1 +1,0 @@
-export { PlainTextCodeBlock } from "@workspace/ui/components/code/plaintext-code";
