+++ packages/ui/package.json
@@ -28,6 +28,9 @@
     "next": "15.3.5",
     "next-themes": "^0.4.6",
     "prettier": "3.6.2",
+    "react-children-utilities": "^2.10.0",
+    "react-markdown": "10.1.0",
+    "remark-gfm": "4.0.1",
     "shiki": "1.27.0",
     "sonner": "2.0.6",
     "tailwind-merge": "^2.6.0",
